requirements.txt
streamlit
google-generativeai
python-dotenv
serpapi
Pillow
google-search-results